discord-forwarder/pyproject.toml
[tool.poetry]
name = "discord-forwarder"
version = "0.1.0"
description = ""
authors = ["lcy <u900011@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"ruamel.yaml" = "*"
requests = "^2.26.0"
pyzmq = "^22.1.0"

[tool.poetry.dev-dependencies]
ipython = "*"
pylint = "*"
pexpect = "*"
mypy = "*"
flake8 = "*"
coverage = "*"
autopep8 = "*"
rope = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
